WEB-4526 move usePrintWindow to own file
