+++ app/pages/clinicworkspace/ClinicPatientsPrintModal/usePrintPDF/usePrintPDF.js
@@ -3,6 +3,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import * as actions from '../../../../redux/actions';
 import buildGenerateAGPImages from './buildGenerateAGPImages';
 import moment from 'moment';
+import usePrintWindow from './usePrintWindow';
 
 import { utils as vizUtils } from '@tidepool/viz';
 const { getTimezoneFromTimePrefs } = vizUtils.datetime;
@@ -16,8 +17,6 @@ import min from 'lodash/min';
 import at from 'lodash/at';
 import map from 'lodash/map';
 import keys from 'lodash/keys';
-import { useTranslation } from 'react-i18next';
-import { useToasts } from '../../../../providers/ToastProvider';
 
 export const STATUS = {
   // States in order of happy path AGP generation sequence
@@ -79,13 +78,11 @@ const inferLastCompletedStep = (patientId, data, pdf, hasPrintStarted) => {
   return STATUS.STATE_CLEARED;
 };
 
-const getFetchLatestDatumPatientOpts = () => {
-  return {
-    initial: true,
-    forceDataWorkerAddDataRequest: true,
-    useCache: false,
-  };
-};
+const getFetchLatestDatumPatientOpts = () => ({
+  initial: true,
+  forceDataWorkerAddDataRequest: true,
+  useCache: false,
+});
 
 const getFetchPatientOpts = (timePrefs, opts) => {
   const enabledOpts = filter(opts, { disabled: false });
@@ -98,70 +95,6 @@ const getFetchPatientOpts = (timePrefs, opts) => {
     forceDataWorkerAddDataRequest: true,
     useCache: false,
   };
-};
-
-const usePrintWindow = () => {
-  const { t } = useTranslation();
-  const { set: setToast } = useToasts();
-
-  const printWindowRef = useRef(null);
-  const getPrintWindow = () => printWindowRef.current;
-
-  // TODO: Clear windowRef to prevent race condition?
-
-  const waitMessage = t('Please wait while Tidepool generates your PDF report.');
-  const waitHTML = `<p align="center" style="margin-top:20px;font-size:16px;font-family:sans-serif">${waitMessage}</p>`;
-
-  const openPrintWindow = () => {
-    if (!printWindowRef.current || printWindowRef.current?.closed) {
-      printWindowRef.current = window.open();
-      printWindowRef.current.document.write(waitHTML);
-    }
-  };
-
-  const triggerPrint = (pdf) => {
-    if (!pdf?.combined?.url) return; // TODO: error?
-
-    if (printWindowRef.current && !printWindowRef.current?.closed) {
-      // If we already have a ref to a PDF window, (re)use it
-      printWindowRef.current.location.href = pdf.combined.url;
-    } else {
-      // Otherwise, we create and open a new PDF window ref.
-      printWindowRef.current = window.open(pdf.combined.url);
-    }
-
-    setTimeout(() => {
-      if (printWindowRef.current) {
-        printWindowRef.current.focus();
-        printWindowRef.current.print();
-      } else {
-        if (!pdf?.combined?.url) return; // TODO: error?
-
-        setToast({
-          message: t('A popup blocker is preventing your report from opening.'),
-          variant: 'warning',
-          autoHideDuration: null,
-          action: (
-            <Button
-              p={0}
-              sx= {{ lineHeight: 1.5, fontSize: 1 }}
-              variant="textPrimary"
-              onClick={() => {
-                printWindowRef.current = window.open(pdf.combined.url);
-                printWindowRef.current.focus();
-                printWindowRef.current.print();
-                setToast(null);
-              }}
-            >
-              {t('Open it anyway')}
-            </Button>
-          ),
-        });
-      }
-    });
-  };
-
-  return { openPrintWindow, triggerPrint };
 };
 
 const usePrintPDF = (
